full mapping model refactor
