+++ src/main/java/com/thedariusz/warnme/twitter/model/Media.java
@@ -1,4 +1,4 @@
-package com.thedariusz.warnme.twitter;
+package com.thedariusz.warnme.twitter.model;
 
 public class Media {
     private String type;
