+++ src/main/java/com/thedariusz/warnme/twitter/model/Meta.java
@@ -1,4 +1,4 @@
-package com.thedariusz.warnme.twitter;
+package com.thedariusz.warnme.twitter.model;
 
 public class Meta {
     private String oldest_id;
